% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{build_F}
\alias{build_F}
\title{Build scoring matrix F}
\usage{
build_F(F, seq_X, seq_Y, match, mismatch, d)
}
\arguments{
\item{F}{Pre-initialized scoring matrix}

\item{seq_X}{A nucleotide sequence preceeded by '0'}

\item{seq_Y}{Another nucleotide sequence preceeded by '0'}

\item{match}{Score associated with a match}

\item{mismatch}{Penalty associated with a mismatch}

\item{d}{Penalty associated with a gap/indel}
}
\value{
Scoring matrix F obtained according to Smith-Waterman algorithm
}
\description{
This function builds and returns the scoring matrix
F according to Smith-Waterman greedy algorithm
}
\references{
\url{https://en.wikipedia.org/wiki/Smith\%E2\%80\%93Waterman_algorithm}\cr
}
\seealso{
\code{\link{smith_waterman_string}}\cr
}
\author{
Emanuele Cavalleri\cr Politecnico di Milano\cr Maintainer: Emanuele
Cavalleri\cr E-Mail: \href{mailto:emanuele.cavalleri@mail.polimi.it}{emanuele.cavalleri@mail.polimi.it}
}
