% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print_smith_waterman}
\alias{print_smith_waterman}
\title{Show Smith-Waterman algorithm output}
\usage{
print_smith_waterman(X, Y, match, mismatch, d)
}
\arguments{
\item{X}{A nucleotidic sequence}

\item{Y}{Another nucleotidic sequence}

\item{match}{Score of a match}

\item{mismatch}{Penalty of a mismatch}

\item{d}{Penalty of a gap/indel}
}
\value{
A visualization of the two aligned sequences with the best score
}
\description{
This function take as input two strings and returns a visualization
of one of the possible best local alignment in terms of score
according to Smith-Waterman greedy algorithm.
}
\examples{
library(Biostrings)
print_smith_waterman(DNAString("TATCT"), "TTCGG", 7, -10, -7)

}
\references{
\url{https://en.wikipedia.org/wiki/Smith\%E2\%80\%93Waterman_algorithm}\cr
}
\seealso{
\code{\link{smith_waterman}}\cr
\code{\link{smith_waterman_DNAString}}\cr
}
\author{
Emanuele Cavalleri\cr Politecnico di Milano\cr Maintainer: Emanuele
Cavalleri\cr E-Mail: \href{mailto:emanuele.cavalleri@mail.polimi.it}{emanuele.cavalleri@mail.polimi.it}
}
