% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{class}
\name{Smith_Waterman-class}
\alias{Smith_Waterman-class}
\alias{Smith_Waterman}
\title{Create Smith_Waterman (RefClass) object}
\arguments{
\item{X}{A nucleotide sequence}

\item{Y}{Another nucleotide sequence}

\item{match}{Score associated with a match}

\item{mismatch}{Penalty associated with a mismatch}

\item{d}{Penalty associated with a gap/indel}

\item{aligned_X}{The aligned nucleotide sequence obtained according to Smith-Waterman algorithm}

\item{aligned_Y}{The other aligned nucleotide sequence obtained according to Smith-Waterman algorithm}

\item{alignment_score}{Best alignment score obtained according to Smith-Waterman algorithm}

\item{F}{Scoring matrix obtained according to Smith-Waterman algorithm}
}
\value{
A new Smith_Waterman object
}
\description{
Create Smith_Waterman (RefClass) object
}

\references{
\url{https://en.wikipedia.org/wiki/Smith\%E2\%80\%93Waterman_algorithm}\cr
}
\author{
Emanuele Cavalleri\cr Politecnico di Milano\cr Maintainer: Emanuele
Cavalleri\cr E-Mail: \href{mailto:emanuele.cavalleri@mail.polimi.it}{emanuele.cavalleri@mail.polimi.it}
}
