% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{smith_waterman}
\alias{smith_waterman}
\title{Apply Smith-Waterman algorithm and show its output}
\usage{
smith_waterman(X, Y, match, mismatch, d)
}
\arguments{
\item{X}{A nucleotide sequence or a DNAString object}

\item{Y}{Another nucleotide sequence or another DNAString object}

\item{match}{Score associated with a match}

\item{mismatch}{Penalty associated with a mismatch}

\item{d}{Penalty associated with a gap/indel}
}
\value{
A Smith-Waterman object
}
\description{
This function takes as input one string and one DNAString object
or two strings or two DNAString objects and returns
one of the possible best local alignment in terms
of score according to Smith-Waterman greedy algorithm.
}
\examples{
library(Biostrings)
smith_waterman(DNAString("TATCT"), "TTCGG", 7, -10, -7)

}
\references{
\url{https://en.wikipedia.org/wiki/Smith\%E2\%80\%93Waterman_algorithm}\cr
}
\seealso{
\code{\link{smith_waterman_string}}\cr
\code{\link{smith_waterman_DNAString}}\cr
}
\author{
Emanuele Cavalleri\cr Politecnico di Milano\cr Maintainer: Emanuele
Cavalleri\cr E-Mail: \href{mailto:emanuele.cavalleri@mail.polimi.it}{emanuele.cavalleri@mail.polimi.it}
}
