% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{smith_waterman_DNAString}
\alias{smith_waterman_DNAString}
\title{Apply Smith-Waterman algorithm to DNAString objects}
\usage{
smith_waterman_DNAString(X, Y, match, mismatch, d)
}
\arguments{
\item{X}{A DNAString object}

\item{Y}{Another DNAString object}

\item{match}{Score associated with a match}

\item{mismatch}{Penalty associated with a mismatch}

\item{d}{Penalty associated with a gap/indel}
}
\value{
A Smith_Waterman object
}
\description{
This function takes as input two DNAString objects
and returns one of the possible best local alignment in terms
of score according to Smith-Waterman greedy algorithm.
}
\examples{
library(Biostrings)
smith_waterman_DNAString(DNAString("TATCT"), DNAString("TTCGG"), 7, -10, -7)

}
\references{
\url{https://en.wikipedia.org/wiki/Smith\%E2\%80\%93Waterman_algorithm}\cr
}
\seealso{
\code{\link{smith_waterman_string}}\cr
\code{\link{smith_waterman}}\cr
}
\author{
Emanuele Cavalleri\cr Politecnico di Milano\cr Maintainer: Emanuele
Cavalleri\cr E-Mail: \href{mailto:emanuele.cavalleri@mail.polimi.it}{emanuele.cavalleri@mail.polimi.it}
}
