% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smithWaterman-package.R
\docType{package}
\name{smithWaterman-package}
\alias{smithWaterman-package}
\alias{smithWaterman}
\title{smithWaterman - Best local alignment of two nucleotide sequences according to Smith-Waterman algorithm}
\description{
The \code{smithWaterman} package implements functions for the local alignment
of two nucleotide sequences according to Smith-Waterman greedy algorithm.
}
\details{
\tabular{ll}{
Package: \tab smithWaterman\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2022-12-19\cr
License: \tab GPL (>=2)\cr
}
}
\references{
\url{https://github.com/emanuelecavalleri/Smith-Waterman}\cr
}
\author{
Emanuele Cavalleri [aut, cre]\cr
Politecnico di Milano\cr
Maintainer: Emanuele Cavalleri\cr
E-Mail: \href{mailto:emanuele.cavalleri@mail.polimi.it}{emanuele.cavalleri@mail.polimi.it}
}
