% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{traceback}
\alias{traceback}
\title{Traceback}
\usage{
traceback(F, seq_X, seq_Y, match, mismatch, d)
}
\arguments{
\item{F}{Scoring matrix obtained according to Smith-Waterman algorithm}

\item{seq_X}{A nucleotide sequence preceded by '0'}

\item{seq_Y}{Another nucleotide sequence preceded by '0'}

\item{match}{Score associated with a match}

\item{mismatch}{Penalty associated with a mismatch}

\item{d}{Penalty associated with a gap/indel}
}
\value{
A list containing the two locally aligned strings
}
\description{
Traceback implemented according to an efficient modified version
of the Needleman-Wunsch algorithm (see Wikipedia reference link).
This function returns a list containing the two locally aligned strings
}
\references{
\url{https://en.wikipedia.org/wiki/Needleman\%E2\%80\%93Wunsch_algorithm#Advanced_presentation_of_algorithm}\cr
}
\seealso{
\code{\link{smith_waterman_string}}\cr
}
\author{
Emanuele Cavalleri\cr Politecnico di Milano\cr Maintainer: Emanuele
Cavalleri\cr E-Mail: \href{mailto:emanuele.cavalleri@mail.polimi.it}{emanuele.cavalleri@mail.polimi.it}
}
