% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{smith_waterman_DNAString}
\alias{smith_waterman_DNAString}
\title{Apply Smith-Waterman algorithm with DNAStrings}
\usage{
smith_waterman_DNAString(X, Y, match, mismatch, d)
}
\arguments{
\item{X}{A DNAString object}

\item{Y}{Another DNAString object}

\item{match}{Score of a match}

\item{mismatch}{Penalty of a mismatch}

\item{d}{Penalty of a gap/indel}
}
\value{
A list containing a possible local alignment with cost matrix and best score
}
\description{
This function take as input two DNAString(Set) (Biostrings) objects
and returns one of the possible best local alignment in terms
of score according to Smith-Waterman greedy algorithm.
}
\examples{
library(Biostrings)
smith_waterman_DNAString(DNAString("TATCT"), DNAString("TTCGG"), 7, -10, -7)

}
\references{
\url{https://en.wikipedia.org/wiki/Smith\%E2\%80\%93Waterman_algorithm}\cr
}
\seealso{
\code{\link{smith_waterman}}\cr
\code{\link{print_smith_waterman}}\cr
}
\author{
Emanuele Cavalleri\cr Politecnico di Milano\cr Maintainer: Emanuele
Cavalleri\cr E-Mail: \href{mailto:emanuele.cavalleri@mail.polimi.it}{emanuele.cavalleri@mail.polimi.it}
}
